#505 refactor fiscal tests
